Module Cobol_ptree.Dummies

Defines some dummy nodes that can be used to fill in missing parts of the parse tree.

val integer_zero : string

val integer_one : string

val alphanum__ : string

val dummy_string : pos:Stdlib.Lexing.position -> string Cobol_common__Srcloc.TYPES.with_loc

val dummy_name : pos:Stdlib.Lexing.position -> string Cobol_common__Srcloc.TYPES.with_loc

val dummy_qualname : pos:Stdlib.Lexing.position -> _ Cobol_ptree__.Terms.term

val dummy_qualname' : pos:Stdlib.Lexing.position -> _ Cobol_ptree__.Terms.term Cobol_common.Srcloc.with_loc

val dummy_qualident : pos:Stdlib.Lexing.position -> Cobol_ptree__.Terms.qualident

val dummy_ident : pos:Stdlib.Lexing.position -> _ Cobol_ptree__.Terms.term

val dummy_literal : _ Cobol_ptree__.Terms.term

val dummy_alphanum : Cobol_ptree__.Terms.alphanum

val dummy_expr : Cobol_ptree__.Terms.expression

val dummy_picture : pos:Stdlib.Lexing.position -> Cobol_ptree__.Data_sections.picture_clause

val dummy_picture_locale : Cobol_ptree__.Data_descr.locale_phrase

val fixed_zero : Cobol_ptree__.Numericals.fixed

val floating_zero : Cobol_ptree__.Numericals.floating

val boolean_zero : Cobol_ptree__.Numericals.boolean

val is_dummy_name : string Cobol_common__Srcloc.TYPES.with_loc -> bool

val is_dummy_qualname : _ Cobol_ptree__.Terms.term -> bool

val strip_dummies_from_qualname : [ Cobol_ptree__.Terms.name_ | Cobol_ptree__.Terms.name Cobol_ptree__.Terms.qual_ ] Cobol_ptree__.Terms.term -> [ Cobol_ptree__.Terms.name_ | Cobol_ptree__.Terms.name Cobol_ptree__.Terms.qual_ ] Cobol_ptree__.Terms.term